+++ assignment4/unitconv/models.py
@@ -1,9 +1,9 @@
 from django.db import models
 
 class Factor(models.Model):
-    t_oz = models.IntegerField()
-    lb = models.IntegerField()
-    oz = models.IntegerField()
-    ton = models.IntegerField()
-    kg = models.IntegerField()
-    g = models.IntegerField()
+    t_oz = models.FloatField()
+    lb = models.FloatField()
+    oz = models.FloatField()
+    ton = models.FloatField()
+    kg = models.FloatField()
+    g = models.FloatField()
